respaldos/monkeycomputer/sevenbarber-app/requirements.txt
streamlit
streamlit-option-menu
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
